- Merge dei due branch ("features/supply" e "features/superadmin-supply"); - Fix vari sui componenti dopo il merge; - Modifica tipo di colonna per i prezzi da float a decimal;

+++ database/migrations/2023_10_24_135421_create_supplies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->text('uuid');
             $table->foreignIdFor(\App\Models\User::class, 'user_id');
-            $table->float('amount');
+            $table->decimal('amount');
             $table->string('payment_method')->nullable();
             $table->enum('status', array_keys(\App\Models\Supply::STATUSES));
             $table->boolean('confirmed')->default(false);
